requirements.txt
opencv-python==4.6.0.66
image-similarity-measures==0.3.5
scikit-learn==1.1.2

numpy~=1.23.4
pytesseract~=0.3.10
pillow~=9.3.0